(refactor) Refactor components using Redux
