+++ src/reducers/reducer_sentiments.js
@@ -1,7 +1,7 @@
 export default function(state = [], action) {
   switch(action.type) {
     case 'SENTIMENT_RESPONSE':
-      return [action.payload];
+      return action.payload;
     default:
       return state;
   }
